+++ src/layouts/ModelizerDrawLayout.vue
@@ -7,14 +7,13 @@
     <modelizer-draw-left-drawer
       v-if="data.plugin"
       :plugin="data.plugin"
-      :templates="templates"
       :project-name="projectName"
     />
 
     <q-page-container>
       <q-splitter
         v-model="splitter"
-        :limits="[50, 100]"
+        :limits="[25, 100]"
         separator-class="separator-class"
         :class="isVisible ? '' : 'splitter-invisible'"
         :style="{ height: `calc(100vh - ${reservedHeight + 70}px)` }"
@@ -59,7 +58,6 @@ import ModelizerDrawLeftDrawer from 'src/components/drawer/ModelizerDrawLeftDraw
 import ComponentDetailPanel from 'src/components/drawer/ComponentDetailPanel.vue';
 import ModelizerDrawPage from 'src/pages/ModelizerDrawPage.vue';
 import { getPluginByName, initComponents } from 'src/composables/PluginManager';
-import { getTemplatesByType } from 'src/composables/TemplateManager';
 import {
   computed,
   onMounted,
@@ -67,9 +65,7 @@ import {
   reactive,
   ref,
 } from 'vue';
-import { useI18n } from 'vue-i18n';
 import { useRoute } from 'vue-router';
-import { Notify } from 'quasar';
 import GitAuthenticationDialog from 'components/dialog/GitAuthenticationDialog.vue';
 import GitAddRemoteDialog from 'components/dialog/GitAddRemoteDialog.vue';
 import ComponentMenuDialog from 'components/dialog/ComponentMenuDialog.vue';
@@ -79,7 +75,6 @@ import DrawerEvent from 'src/composables/events/DrawerEvent';
 import AIChatDrawer from 'components/drawer/AIChatDrawer.vue';
 import DeleteAIConversationDialog from 'components/dialog/DeleteAIConversationDialog.vue';
 
-const { t } = useI18n();
 const route = useRoute();
 
 const query = computed(() => route.query);
@@ -89,7 +84,6 @@ const pluginName = computed(() => query.value.plugin);
 const splitter = ref(100);
 const splitterKey = ref('');
 const isVisible = ref(false);
-const templates = ref([]);
 const componentId = ref('');
 const reservedHeight = ref(37);
 
@@ -122,18 +116,6 @@ async function initView() {
       })));
       data.plugin.draw();
     }),
-    getTemplatesByType(
-      'component',
-      data.plugin.data.name,
-    ).then((response) => {
-      templates.value = response;
-    }).catch(() => {
-      Notify.create({
-        type: 'negative',
-        message: t('errors.templates.getData'),
-        html: true,
-      });
-    }),
   ]);
 }
 
@@ -152,7 +134,7 @@ function onDrawerEvent({ key, type, id }) {
   componentId.value = id || null;
   splitterKey.value = key;
   isVisible.value = type === 'open';
-  splitter.value = type === 'open' ? 75 : 100;
+  splitter.value = type === 'open' ? 60 : 100;
 }
 
 onMounted(() => {